Admin Fixes

+++ controllers/user.js
@@ -32,18 +32,23 @@ exports.get = (req, res) => {
 // Update a specific user
 exports.put = (req, res) => {
 	const data = req.body.data
-	User.findByIdAndUpdate({ _id: data._id }, {name: data.name, username: data.username}, { new: true })
-	.then(user => {
-		if (!user) {
-			return res.sendStatus(404);
-		}
-		user.password = undefined;
-		user.recoveryCode = undefined;
-		res.json(user);
-	})
-	.catch(err => {
-		res.status(422).send(err.errors);
-	});
+	console.log(req.user._id + ' trying to update user:' + data._id)
+	if ((req.user.admin === true) | ((req.user._id).toString() === data._id)){ //Admin check works, todo: check for role/memberships in the future
+		User.findByIdAndUpdate({ _id: data._id }, {name: data.name, username: data.username}, { new: true })
+			.then(user => {
+				if (!user) {
+					return res.sendStatus(404);
+				}
+				user.password = undefined;
+				user.recoveryCode = undefined;
+				res.json(user);
+			})
+			.catch(err => {
+				res.status(422).send(err.errors);
+			});
+	} else {
+		res.status(401).send({'message':'You do not have permission to complete this action'})
+	}
 };
 
 // Create a new user
@@ -61,16 +66,20 @@ exports.post = (req, res) => {
 
 // Remove a user record TODO: set 'active' flag rather than actually delete the user object.
 exports.delete = (req, res) => {
-	User.findByIdAndRemove(
-		{ _id: req.body._id }
-	)
-	.then(user => {
-		if (!user) {
-			return res.sendStatus(404);
-		}
-		res.sendStatus(204);
-	})
-	.catch(err => {
-		res.status(422).send(err.errors);
-	});
+	if ((req.user.admin === true) | ((req.user._id).toString() === data._id)){ //Admin check works, todo: check for role/memberships in the future
+		User.findByIdAndRemove(
+			{ _id: req.body._id }
+		)
+			.then(user => {
+				if (!user) {
+					return res.sendStatus(404);
+				}
+				res.sendStatus(204);
+			})
+			.catch(err => {
+				res.status(422).send(err.errors);
+			});
+	} else {
+		res.status(401).send({'message':'You do not have permission to complete this action'})
+	}
 };
